// src/generated/responses.ts
export interface ReleasesClass {
  totalCount: number;
  items: ItemClass[];
}
// ...
export interface AlbumCopyright {
  totalCount: number;
  items: BiographyElement[];
}

export interface BiographyElement {
  type: BiographyType;
  text: string;
}

export enum BiographyType {
  Autobiography = 'AUTOBIOGRAPHY',
  C = 'C',
  P = 'P',
}

export interface CoverArt {
  extractedColors: PurpleExtractedColors;
  sources: Icon[];
}

export interface PurpleExtractedColors {
  colorRaw: Color;
  colorLight: Color;
  colorDark: Color;
}

export interface Color {
  hex: string;
}

export interface ReleaseDateClass {
  isoString: Date;
  precision: Precision;
}

export enum Precision {
  Day = 'DAY',
  Minute = 'MINUTE',
}

export interface Discs {
  totalCount: number;
  items: DiscsItem[];
}
// ...
export interface LatestPlayability {
  playable: boolean;
  reason: PurpleReason;
}

export enum PurpleReason {
  CountryRestricted = 'COUNTRY_RESTRICTED',
  Playable = 'PLAYABLE',
}

export interface AlbumSharingInfo {
  shareId: string;
  shareUrl: string;
}

export enum LatestType {
  Album = 'ALBUM',
  Compilation = 'COMPILATION',
  Ep = 'EP',
  Single = 'SINGLE',
}

export interface PurpleTracks {
  totalCount: number;
  items: IndigoItem[];
}

export interface IndigoItem {
  track: PurpleTrack;
}

export interface PurpleTrack {
  playability: LatestPlayability;
  duration: TrackDurationClass;
}

export interface TrackDurationClass {
  totalMilliseconds: number;
}
// ...
export interface ContentRating {
  label: Label;
}
// ...
export interface ImagesClass {
  items: AvatarImageElement[];
}
// ...
export interface ArtistStats {
  followers: number;
  monthlyListeners: number;
  worldRank: number;
  topCities: TopCities;
}
// ...
export enum ArtistUnionTypename {
  Album = 'Album',
  Artist = 'Artist',
  Playlist = 'Playlist',
  Track = 'Track',
  User = 'User',
}

export interface IndigoDiscography {
  latest: LatestElement;
  popularReleasesAlbums: PopularReleasesAlbums;
  singles: CompilationsClass;
  albums: CompilationsClass;
  compilations: CompilationsClass;
  topTracks: PurpleTopTracks;
}
// ...
export interface ArtistUnionRelatedContent {
  appearsOn: PurpleAppearsOn;
  featuringV2: V2;
  discoveredOnV2: V2;
  relatedArtists: RelatedArtists;
}
// ...
export interface GetPartnerAlbumData {
  albumUnion: AlbumUnion;
}

export interface AlbumUnion {
  __typename: ArtistUnionTypename;
  copyright: AlbumCopyright;
  courtesyLine: string;
  date: ReleaseDateClass;
  isPreRelease: boolean;
  label: string;
  name: string;
  playability: LatestPlayability;
  preReleaseEndDateTime: null;
  sharingInfo: AlbumSharingInfo;
  tracksV2: TracksV2;
  type: LatestType;
  uri: string;
  visualIdentity: AlbumUnionVisualIdentity;
  watchFeedEntrypoint: WatchFeedEntrypoint;
  artists: ReleasesClass;
  coverArt: CoverArt;
  discs: Discs;
  releases: ReleasesClass;
  moreAlbumsByArtist: AlbumUnionMoreAlbumsByArtist;
}

export interface AlbumUnionMoreAlbumsByArtist {
  items: Item5[];
}

export interface Item5 {
  discography: HilariousDiscography;
}

export interface HilariousDiscography {
  popularReleasesAlbums: PurplePopularReleasesAlbums;
}

export interface TracksV2 {
  items: TracksV2Item[];
  totalCount: number;
}
// ...
export interface AlbumUnionVisualIdentity {
  squareCoverImage: SquareCoverImageClass;
}
// ...
export interface WatchFeedEntrypoint {
  entrypointUri: string;
  thumbnailImage: ThumbnailImage;
  video: Video | null;
}
// ...
export interface GetPartnerArtistDiscographyData {
  artistUnion: GetPartnerArtistDiscographyDataArtistUnion;
}

export interface GetPartnerArtistDiscographyDataArtistUnion {
  __typename: ArtistUnionTypename;
  discography: AmbitiousDiscography;
}

export interface AmbitiousDiscography {
  all: Singles;
}

export interface GetPartnerArtistOverviewData {
  artistUnion: GetPartnerArtistOverviewDataArtistUnion;
}

export interface GetPartnerArtistOverviewDataArtistUnion {
  __typename: ArtistUnionTypename;
  discography: IndigoDiscography;
  goods: FluffyGoods;
  headerImage: HeaderImage;
  id: string;
  onPlatformReputationTrait: OnPlatformReputationTrait;
  preRelease: null;
  profile: TentacledProfile;
  relatedContent: ArtistUnionRelatedContent;
  relatedMusicVideos: RelatedMusicVideosClass;
  sharingInfo: AlbumSharingInfo;
  stats: ArtistStats;
  unmappedMusicVideosV2: RelatedMusicVideosClass;
  uri: string;
  visualIdentity: ArtistUnionVisualIdentity;
  visuals: TentacledVisuals;
  watchFeedEntrypoint: WatchFeedEntrypoint;
}

export interface FluffyGoods {
  concerts: ReleasesClass;
  merch: FluffyMerch;
}

export interface HeaderImage {
  data: ImageData;
}

export interface ImageData {
  __typename: FluffyTypename;
  sources: Source[];
}

export interface OnPlatformReputationTrait {
  verification: Verification;
}

export interface Verification {
  isRegistered: boolean;
  isVerified: boolean;
}

export interface TentacledProfile {
  biography: BiographyElement;
  externalLinks: ExternalLinks;
  name: LabelEnum;
  pinnedItem: TentacledPinnedItem;
  playlistsV2: PlaylistsV2;
}
// ...
export interface RelatedMusicVideosClass {
  __typename?: string;
  items: RelatedMusicVideosItem[];
  pagingInfo?: AlbumsPagingInfo;
  totalCount: number;
}
// ...
export interface ArtistUnionVisualIdentity {
  wideFullBleedImage: SquareCoverImageClass;
}

export interface TentacledVisuals {
  avatarImage: HeaderImageClass;
  gallery: ImagesClass;
}
// ...
export interface GetPartnerPlaylistData {
  playlistV2: PlaylistV2;
}

export interface PlaylistV2 {
  __typename: ArtistUnionTypename;
  content: Content;
  attributes: Attribute[];
  basePermission: string;
  currentUserCapabilities: CurrentUserCapabilities;
  description: string;
  followers: number;
  following: boolean;
  format: Format;
  images: ImagesClass;
  members: Members;
  name: string;
  ownerV2: OwnerV2;
  revisionId: string;
  sharingInfo: AlbumSharingInfo;
  uri: string;
  visualIdentity: AlbumUnionVisualIdentity;
  watchFeedEntrypoint: WatchFeedEntrypoint;
}

export interface Attribute {
  key: string;
  value: string;
}

export interface Content {
  __typename: string;
  items: ContentItem[];
  pagingInfo: ContentPagingInfo;
  totalCount: number;
}
// ...
export interface TrackUnionAssociationsV3 {
  audioAssociations: AudioAssociations;
  videoAssociations: AlbumsV2;
}

export interface AudioAssociations {
  __typename: AudioAssociationsTypename;
  items: any[];
}

export enum AudioAssociationsTypename {
  TrackAudioAssociationPage = 'TrackAudioAssociationPage',
}

export enum MediaType {
  Audio = 'AUDIO',
  Video = 'VIDEO',
}

export interface ContentPagingInfo {
  limit: number;
  offset: number;
}

export interface CurrentUserCapabilities {
  canAdministratePermissions: boolean;
  canCancelMembership: boolean;
  canEditItems: boolean;
  canView: boolean;
}

export enum Format {
  ArtistMixReader = 'artist-mix-reader',
  Editorial = 'editorial',
  Empty = '',
  FormatShowsShuffle = 'format-shows-shuffle',
  InspiredbyMix = 'inspiredby-mix',
}

export interface Members {
  items: MembersItem[];
  totalCount: number;
}

export interface MembersItem {
  isOwner: boolean;
  permissionLevel: string;
  user: OwnerV2;
}

export interface OwnerV2 {
  data: UserData;
}
// ...
export interface GetPartnerTrackCountData {
  result: string;
  message: string;
  spotify_track_id: string;
  isrc: string;
  streams: number;
}

export interface GetPartnerTrackData {
  trackUnion: TrackUnion;
}

export interface TrackUnion {
  __typename: ArtistUnionTypename;
  associationsV3: TrackUnionAssociationsV3;
  contentRating: ContentRating;
  duration: TrackDurationClass;
  id: string;
  mediaType: MediaType;
  name: string;
  playability: LatestPlayability;
  playcount: string;
  sharingInfo: AlbumSharingInfo;
  trackNumber: number;
  uri: string;
  visualIdentity: AlbumUnionVisualIdentity;
  albumOfTrack: TrackUnionAlbumOfTrack;
  firstArtist: FirstArtist;
  otherArtists: ImagesClass;
}

export interface TrackUnionAlbumOfTrack {
  copyright: AlbumCopyright;
  courtesyLine: string;
  id: string;
  date: AlbumOfTrackDate;
  name: string;
  playability: AlbumOfTrackPlayability;
  sharingInfo: AlbumSharingInfo;
  tracks: AlbumOfTrackTracks;
  type: LatestType;
  uri: string;
  coverArt: HeaderImageClass;
}

export interface AlbumOfTrackTracks {
  items: Item6[];
  totalCount: number;
}

export interface Item6 {
  track: IndigoTrack;
}

export interface IndigoTrack {
  trackNumber: number;
  uri: string;
}

export interface FirstArtist {
  items: FirstArtistItem[];
  totalCount: number;
}